+++ src/main/java/net/unicon/persondir/config/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Classes providing support for person directory Spring application context custom configuration XML schema
+ */
+package net.unicon.persondir.config;
+