+++ changelog.md
@@ -1,6 +1,9 @@
-## Changelog
+## person-directory-addons changelog
 
-###Changes in version 0.1
+###Changes in version 0.2
 ======================================
+
+* added [Custom namespace XML schema](https://github.com/Unicon/person-directory-addons/tree/master/src/main/resources/net/unicon/persondir/config/person-directory.xsd)
+* added [Custom namespace XML schema parsers](https://github.com/Unicon/person-directory-addons/blob/master/src/main/java/net/unicon/persondir/config/PersonDirectoryNamespaceHandler.java)